+++ examples/storybook/stories/Plugins/Audio/AudioViewer.tsx
@@ -1,0 +1,11 @@
+import type { FunctionComponent } from 'react';
+import React from 'react';
+import type { DraftContent } from 'ricos-viewer';
+import { RicosViewer } from 'ricos-viewer';
+import { pluginAudio } from 'wix-rich-content-plugin-audio/dist/module.viewer';
+
+const Viewer: FunctionComponent<{ content?: DraftContent }> = ({ content }) => (
+  <RicosViewer content={content} plugins={[pluginAudio()]} />
+);
+
+export default Viewer;