+++ CHANGELOG.md
@@ -17,6 +17,12 @@
     Click to see more.
   </summary>
 
+### :house: Internal
+- `audio`
+  - [#3724](https://github.com/wix/ricos/pull/3724) add audio to storybook
+### :bug: Bug Fix
+- `audio`
+  - [#3721](https://github.com/wix/ricos/pull/3721) add err msg when upload fails & some bug fixes
 
 
 </details>
@@ -28,7 +34,6 @@
 
 - `audio`
   - [#3718](https://github.com/wix/ricos/pull/3718) fix audio bi &  a11y (clear naming for screen readers)
-  - [#3721](https://github.com/wix/ricos/pull/3721) add err msg when upload fails & some bug fixes
 - `button`
   - [#3719](https://github.com/wix/ricos/pull/3719) fix button settings BM theming